ui_manager/requirements.txt
Flask[async]==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.0.0
pyjwt==2.6.0
mysqlclient==2.1.1
Flask-SQLAlchemy==3.0.3
